msi-generation-service/frontend_app/requirements.txt
streamlit
requests
Pillow
numpy